davinci: Ramp up fastcharge threshold
Xiaomi seems to keep the threshold at 35C while the ambient temps in India are pretty much same as that.
This ends up preventing fastcharge for most of the time. Ramp it up to 39C so that fastcharges are triggered much more consistently.
This does not modify the cutoff temps which are at 45C.

Signed-off-by: baalajimaestro <me@baalajimaestro.me>